backend/pyproject.toml
[tool.poetry]
name = "method_interview"
version = "0.1.0"
description = ""
authors = ["Jake Kearns <jakearns95@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
alembic = "1.7.5"
boto3 = "1.17.71"
botocore = "1.20.71"
fastapi = "0.79.0"
gunicorn = "20.1.0"
jinja2 = "3.0.3"
python-dateutil = "2.8.1"
python-dotenv = "0.19.2"
pyyaml = "6.0.1"
requests = "2.25.1"
typing-extensions = "4.3.0"
urllib3 = "1.25.11"
uvicorn = "0.20.0"
celery = { extras = ["redis"], version = "^5.2.7" }
flower = "^1.2.0"
black = "22.6.0"
isort = "5.10.1"
motor= "2.3.0"
dnspython = "2.4.2"


[tool.black]
exclude = '/(\.direnv|\.eggs|\.git|\.hg|\.mypy_cache|\.nox|\.tox|\.venv|venv|\.svn|\.ipynb_checkpoints|_build|buck-out|build|dist|__pypackages__|node_modules)/'

[tool.isort]
profile = "black"
extend_skip_glob = [".*", "node_modules/*"]